fix(activity): preserve status behavior in sidebar
